fix: resolve 3 lint issues introduced in maintain-disk refactor (#8)
* fix: resolve 3 lint issues introduced in maintain-disk refactor

- findLogsToRotate: drop always-nil error return (unparam)
- maintain_disk.go: check fmt.Scanln return value (errcheck)
- beadwatcher/watcher.go: fix "cancelled" → "canceled" spelling (misspell)


* fix: cross-platform build for FreeBSD (disk_usage_unix.go)

On FreeBSD, syscall.Statfs_t.Bavail is int64 (not uint64 like Linux/Darwin).
The expression stat.Bavail * uint64(stat.Bsize) fails with "mismatched types
int64 and uint64" on FreeBSD.

Fix by explicitly casting stat.Bavail to uint64 before multiplication,
which compiles cleanly on all platforms.


* fix: make TestEnsureBranchCIDispatchesFallback timing-deterministic

The test used PushWait: time.Millisecond expecting 2 run-list polls before
dispatch, but on slow Windows CI the 1ms deadline expires after just one
poll, causing the second step (run-list) to mismatch the actual call
(workflow-run).

Fix: use time.Nanosecond so the deadline is guaranteed to expire after the
first FindRun on any machine, making the script exactly one run-list →
workflow-run → run-list → run-view with no timing dependency.


* fix: make EnsureBranchCI push-wait deadline injectable for deterministic tests

The push-wait loop used time.Now() directly, making TestEnsureBranchCIDispatchesFallback
timing-sensitive. On Windows CI (coarse 15ms timer resolution), the 1ms PushWait
expired after only 1 poll instead of 2, causing the scriptedRunner to see
'workflow run' where it expected 'run list'.

Add EnsureOptions.NowFn (optional clock override) so tests can inject a fake
clock that advances deterministically. Production code path unchanged (defaults
to time.Now).


---------

+++ internal/daemon/disk_usage_unix.go
@@ -14,8 +14,8 @@ func getDiskUsage(path string) (*DiskUsageInfo, error) {
 		return nil, fmt.Errorf("statfs %s: %w", path, err)
 	}
 
-	totalBytes := stat.Blocks * uint64(stat.Bsize)  //nolint:unconvert
-	freeBytes := stat.Bavail * uint64(stat.Bsize)    //nolint:unconvert
+	totalBytes := stat.Blocks * uint64(stat.Bsize)         //nolint:unconvert
+	freeBytes := uint64(stat.Bavail) * uint64(stat.Bsize) //nolint:unconvert
 	usedBytes := totalBytes - freeBytes
 
 	var usedFraction float64
